HARMONY-1900: More refactoring.
