+++ docs/user/notebook.html
@@ -8109,7 +8109,7 @@ root group (NETCDF4 data model, file format HDF5):
     history: Fri May 27 16:34:47 2022: GDAL CreateCopy( out/nc/001_00_7f00ff_global.nc, ... )
     dimensions(sizes): lon(3600), lat(1800)
     variables(dimensions): |S1 crs(), float64 lat(lat), float64 lon(lon), uint8 red_var(lat, lon), uint8 green_var(lat, lon), uint8 blue_var(lat, lon), uint8 alpha_var(lat, lon)
-    groups: 
+    groups:
 </pre>
 </div>
 </div>
@@ -8122,7 +8122,7 @@ root group (NETCDF4 data model, file format HDF5):
 </div>
 <div class="jp-InputArea jp-Cell-inputArea"><div class="jp-InputPrompt jp-InputArea-prompt">
 </div><div class="jp-RenderedHTMLCommon jp-RenderedMarkdown jp-MarkdownOutput" data-mime-type="text/markdown">
-<p>Example of submitting a request with WKT spatial. The supported WKT geometry types are listed at: <a href="https://harmony-py.readthedocs.io/en/latest/api.html#harmony.harmony.WKT">https://harmony-py.readthedocs.io/en/latest/api.html#harmony.harmony.WKT</a></p>
+<p>Example of submitting a request with WKT spatial. The supported WKT geometry types are listed at: <a href="https://harmony-py.readthedocs.io/en/latest/api.html#harmony.client.WKT">https://harmony-py.readthedocs.io/en/latest/api.html#harmony.client.WKT</a></p>
 </div>
 </div>
 </div>
